Fix bug related to multi-selection operations  sss
The bug caused objects to move unpredictably when
moving multiple selected objects at once.

This happened because the object's position in the
grid was updated sequentially, as such the object
that followed would get assigned the incorrect
position in the grid. Consequently 'Select' would
return the object that had been moved to the
new position, and not the actual selected object.