Add Canvas version of palette entity preview
This is now the default way of showing the
current palette entity. This change improves
scaling different screen sizes and gives a small
improvent on the time it takes for the overlay to
update.

The old grid renderer is still available by
setting PreviewPosition to Center. The grid uses
GameObjects like the PaletteDisplay screen, which
means multiple previews can be diplayed next to
each other. This comes at the cost of obstructing
some screen space, so ultimately this is left as
an alternative for the user.