Improve Cursors performance
Allow cursors to be added without checking for
duplicates in the backing dictionary. This is
useful when we can guarantee no duplicates will
be created while modifying the cursors list.

Cursors will still need to be synced after using
the unchecked methods, but this may be called
once instead of during every call to
Buffer.Transform.

Measured ~10x faster Buffer.SelectInRect calls in
a 100x100 grid.